requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
Authlib==1.3.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.0
google-api-core==2.24.2
google-auth==2.35.0
google-cloud-speech==2.32.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mysql-connector-python==9.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.4.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.38.6
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
